Move AppNexus bid-config to own module, overhaul tests
