+++ static/src/javascripts/projects/commercial/modules/prebid/bid-config.js
@@ -31,7 +31,6 @@ import type {
     PrebidXaxisParams,
 } from './types';
 import {
-    getLargestSize,
     containsBillboard,
     containsDmpu,
     containsLeaderboard,
@@ -50,6 +49,7 @@ import {
     isInAuRegion,
     stripDfpAdPrefixFrom,
 } from './utils';
+import { getAppNexusPlacementId, getAppNexusDirectBidParams } from './appnexus';
 
 const isInSafeframeTestVariant = (): boolean => {
     const variant = getVariant(commercialPrebidSafeframe, 'variant');
@@ -216,100 +216,6 @@ const getImprovePlacementId = (sizes: PrebidSize[]): number => {
             default:
                 return -1;
         }
-    }
-};
-
-const getAppNexusInvCode = (sizes: Array<PrebidSize>): ?string => {
-    const device: string = getBreakpointKey();
-    const section: string = config.get('page.section', 'unknown');
-    const slotSize: PrebidSize | null = getLargestSize(sizes);
-    if (slotSize) {
-        return `${device}${section.toLowerCase()}${slotSize.join('x')}`;
-    }
-};
-
-const getAppNexusDirectPlacementId = (
-    sizes: PrebidSize[],
-    isAuRegion: boolean
-): string => {
-    if (isAuRegion) {
-        return '11016434';
-    }
-
-    const defaultPlacementId: string = '9251752';
-    switch (getBreakpointKey()) {
-        case 'D':
-            if (containsMpuOrDmpu(sizes)) {
-                return '9251752';
-            }
-            if (containsLeaderboardOrBillboard(sizes)) {
-                return '9926678';
-            }
-            return defaultPlacementId;
-        case 'M':
-            if (containsMpu(sizes)) {
-                return '4298191';
-            }
-            return defaultPlacementId;
-        case 'T':
-            if (containsMpu(sizes)) {
-                return '11600568';
-            }
-            if (containsLeaderboard(sizes)) {
-                return '11600778';
-            }
-            return defaultPlacementId;
-        default:
-            return defaultPlacementId;
-    }
-};
-
-const getAppNexusDirectBidParams = (
-    sizes: PrebidSize[],
-    isAuRegion: boolean
-): PrebidAppNexusParams => {
-    if (isAuRegion && config.get('switches.prebidAppnexusInvcode', false)) {
-        return {
-            invCode: getAppNexusInvCode(sizes) || '',
-            member: '7012',
-            keywords: buildAppNexusTargetingObject(buildPageTargeting()),
-        };
-    }
-    return {
-        placementId: getAppNexusDirectPlacementId(sizes, isAuRegion),
-        keywords: buildAppNexusTargetingObject(buildPageTargeting()),
-    };
-};
-
-const getAppNexusPlacementId = (sizes: PrebidSize[]): string => {
-    const defaultPlacementId: string = '13915593';
-    if (isInUsRegion() || isInAuRegion()) {
-        return defaultPlacementId;
-    }
-    switch (getBreakpointKey()) {
-        case 'D':
-            if (containsMpuOrDmpu(sizes)) {
-                return '13366606';
-            }
-            if (containsLeaderboardOrBillboard(sizes)) {
-                return '13366615';
-            }
-            return defaultPlacementId;
-        case 'M':
-            if (containsMpu(sizes)) {
-                return '13366904';
-            }
-            return defaultPlacementId;
-        case 'T':
-            if (containsMpu(sizes)) {
-                return '13366913';
-            }
-            if (containsLeaderboard(sizes)) {
-                return '13366916';
-            }
-            return defaultPlacementId;
-        default:
-            return defaultPlacementId;
     }
 };
 
@@ -693,9 +599,6 @@ export const bids: (string, PrebidSize[]) => PrebidBid[] = (
 
 export const _ = {
     getAdYouLikePlacementId,
-    getAppNexusInvCode,
-    getAppNexusDirectBidParams,
-    getAppNexusPlacementId,
     getDummyServerSideBidders,
     getIndexSiteId,
     getImprovePlacementId,
